requirements.txt
alembic==1.10.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Markdown==0.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
pycparser==2.21
python-dotenv==1.0.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
Werkzeug==2.2.2
